fix: rename documentHistory-restore to use underscore for Convex compatibility
Convex module paths cannot contain hyphens. Renamed to documentHistory_restore.ts
and updated all imports.
